requirements.txt
pandas
psycopg2-binary
pymongo
matplotlib
seaborn
scikit-learn
textblob
nltk
